code review item, new module for bulk request logic
