+++ test/support/test_model.ex
@@ -62,11 +62,11 @@ defimpl ExlasticSearch.Indexable,
   for: [ExlasticSearch.TestModel, ExlasticSearch.MultiVersionTestModel] do
   def id(%{id: id}), do: id
 
-  def document(struct) do
+  def document(struct, _) do
     struct
     |> Map.from_struct()
     |> Map.take(@for.__mappings__())
   end
 
-  def preload(struct), do: struct
+  def preload(struct, _), do: struct
 end
